TH-209 Feat: Add early return logic for checking forbidden in issue

+++ v2/src/ctrlfbe/views.py
@@ -246,13 +246,16 @@ class IssueViewSet(CtrlfAuthenticationMixin, ModelViewSet):
 class IssueDeleteView(CtrlfAuthenticationMixin, APIView):
     @swagger_auto_schema(**SWAGGER_ISSUE_DELETE_VIEW)
     def delete(self, request, *args, **kwargs):
-        self._ctrlf_authentication(request)
+        user = self._ctrlf_authentication(request)
         issue = Issue.objects.filter(id=request.data["issue_id"]).first()
         if issue is None:
             return Response(data={"message": "이슈 ID를 찾을 수 없습니다."}, status=status.HTTP_404_NOT_FOUND)
 
         if issue.status == CtrlfIssueStatus.APPROVED:
             return Response(data={"message": "유효한 요청이 아닙니다"}, status=status.HTTP_400_BAD_REQUEST)
+
+        if issue.owner != user:
+            return Response(data={"message": "권한이 없습니다"}, status=status.HTTP_403_FORBIDDEN)
 
         issue.delete()
         return Response(data={"message": "이슈 삭제"}, status=status.HTTP_204_NO_CONTENT)
